backend/ml_advanced/requirements.txt
xgboost
pandas
numpy
flask
flask-cors
scikit-learn
joblib
pymongo
python-dotenv
